chore: relocate components/auth files to their respective new locations
