+++ apps/web/app/(auth)/auth/forgot-password/reset/page.tsx
@@ -1,5 +1,5 @@
-import { ResetPasswordForm } from "@/app/components/auth/ResetPasswordForm";
-import FormWrapper from "@/app/components/auth/FormWrapper";
+import FormWrapper from "@/app/(auth)/auth/components/FormWrapper";
+import { ResetPasswordForm } from "@/app/(auth)/auth/forgot-password/reset/components/ResetPasswordForm";
 
 const ResetPasswordPage: React.FC = () => {
   return (
