+++ apps/web/app/(auth)/auth/signup-without-verification-success/page.tsx
@@ -1,5 +1,5 @@
-import BackToLoginButton from "@/app/components/auth/BackToLoginButton";
-import FormWrapper from "@/app/components/auth/FormWrapper";
+import BackToLoginButton from "@/app/(auth)/auth/components/BackToLoginButton";
+import FormWrapper from "@/app/(auth)/auth/components/FormWrapper";
 
 export default function SignupWithoutVerificationSuccess() {
   return (
