+++ apps/web/app/(auth)/auth/verification-requested/components/RequestVerificationEmail/index.tsx
