+++ apps/web/app/(auth)/auth/verify/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import { SignIn } from "@/app/components/auth/SignIn";
 import { useSearchParams } from "next/navigation";
-import FormWrapper from "@/app/components/auth/FormWrapper";
+import { SignIn } from "@/app/(auth)/auth/verify/components/SignIn";
+import FormWrapper from "@/app/(auth)/auth/components/FormWrapper";
 
 export default function Verify() {
   const searchParams = useSearchParams();
